+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/ScheduledRelease.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.accountinfo;
 
+import com.concordium.sdk.types.Timestamp;
 import com.concordium.sdk.transactions.Hash;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -11,12 +12,12 @@ import java.util.List;
 @Getter
 @ToString
 public final class ScheduledRelease {
-    private final double releaseTimestamp;
+    private final Timestamp releaseTimestamp;
     private final String releaseAmount;
     private final List<Hash> releaseTransactions;
 
     @JsonCreator
-    ScheduledRelease(@JsonProperty("timestamp") double releaseTimestamp,
+    ScheduledRelease(@JsonProperty("timestamp") Timestamp releaseTimestamp,
                      @JsonProperty("amount") String releaseAmount,
                      @JsonProperty("transactions") List<Hash> releaseTransactions) {
         this.releaseTimestamp = releaseTimestamp;