+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/ScheduledRelease.java
@@ -1,5 +1,7 @@
 package com.concordium.sdk.responses.accountinfo;
 
+import com.concordium.sdk.types.Timestamp;
+import com.concordium.sdk.transactions.Hash;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -10,14 +12,14 @@ import java.util.List;
 @Getter
 @ToString
 public final class ScheduledRelease {
-    private final int releaseTimestamp;
+    private final Timestamp releaseTimestamp;
     private final String releaseAmount;
-    private final List<String> releaseTransactions;
+    private final List<Hash> releaseTransactions;
 
     @JsonCreator
-    ScheduledRelease(@JsonProperty("releaseTimestamp") int releaseTimestamp,
-                     @JsonProperty("releaseAmount") String releaseAmount,
-                     @JsonProperty("releaseTransactions") List<String> releaseTransactions) {
+    ScheduledRelease(@JsonProperty("timestamp") Timestamp releaseTimestamp,
+                     @JsonProperty("amount") String releaseAmount,
+                     @JsonProperty("transactions") List<Hash> releaseTransactions) {
         this.releaseTimestamp = releaseTimestamp;
         this.releaseAmount = releaseAmount;
         this.releaseTransactions = releaseTransactions;