+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased changes
+- Fix parsing bug for `AccountInfo`
+- Transaction `Status` now public
+- `BlockItem` now public
 
 ## 1.3.0
 - Support for transaction deserialization.